Oracles

Oracles

In transition-based dependency parsing, oracles are used to map parser configurations (intermediate parser states) to gold transitions.

Static and dynamic oracles can both be used to iterate through trees.

julia> oracle = DynamicOracle(ArcEager())
julia> for state in oracle(gold_tree)
           # ...
       end

state here will be a DependencyTrees.OracleState, which keeps track of the current configuration, possible transitions, and gold transitions.

DependencyTrees.OracleState — Type.

OracleState

Temporary state for building a transition parse.

source

The iterator itself is something something

DependencyTrees.TreeOracle — Type.

TreeOracle

A gold tree for training.

Collects an oracle, a gold tree, and an exploration policy.

source

DependencyTrees.UnparsableTree — Type.

UnparsableTree

Represents trees that an oracle can't parse.

source

Both kinds of oracles can build parse trees with labelled (typed) or unlabelled (untyped) dependency relations using the arc keyword argument.

DependencyTrees.untyped — Function.

untyped(token)

Create an arc without a dependency label.

source

DependencyTrees.typed — Function.

typed(token)

Create an arc with a labeled dependency relation.

source

Static Oracles

DependencyTrees.StaticOracle — Type.

StaticOracle

Static (deterministic) oracle for mapping parser configurations to gold transitions.

source

Static Oracle Functions

DependencyTrees.static_oracle — Function.

static_oracle(cfg, gold_tree)

Static oracle for arc-standard dependency parsing. Closes over gold trees, mapping parser configurations to optimal transitions.

source

static_oracle(cfg, tree, transition)

Default static oracle function for arc-eager dependency parsing.

See Goldberg & Nivre 2012. (Also called Arc-Eager-Reduce in Qi & Manning 2017).

source

static_oracle(::ArcHybrid, tree)

Static oracle for arc-hybrid dependency parsing. Closes over gold trees, mapping parser configurations to optimal transitions.

source

static_oracle(::ArcSwiftConfig, graph)

Return a training oracle function which returns gold transition operations from a parser configuration with reference to graph.

Described in Qi & Manning 2017.

source

static_oracle(::ListBasedNonProjectiveConfig, tree)

Return a training oracle function which returns gold transition operations from a parser configuration with reference to graph.

source

DependencyTrees.static_oracle_prefer_shift — Function.

static_oracle_prefer_shift(cfg, gold_tree, arc=untyped)

Static oracle for arc-eager dependency parsing. Similar to the "regular" static oracle, but always Shift when ambiguity is present.

See Qi & Manning 2017.

source

Dynamic Oracles

DependencyTrees.DynamicOracle — Type.

DynamicOracle(system, oracle_function = dynamic_oracle; arc = untyped)

Create a dynamic oracle for dependency parsing (see Goldberg & Nivre, 2012).

Dynamic oracles allow multiple gold transitions from each parser configuration.

source

Dynamic Oracle Functions

DependencyTrees.dynamic_oracle — Function.

dynamic_oracle(t, cfg::ArgEagerConfig, tree)

Dynamic oracle function for arc-eager parsing.

For details, see Goldberg & Nivre 2012.

source

dynamic_oracle(t, cfg::ArgHybridConfig, tree)

Dynamic oracle function for arc-hybrid parsing.

For details, see Goldberg & Nivre, 2013.

source

Exploration Policies

Something about exploration during training.

DependencyTrees.AlwaysExplore — Type.

AlwaysExplore()

Policy for always exploring sub-optimal transitions.

source

DependencyTrees.NeverExplore — Type.

NeverExplore()

Policy for never exploring sub-optimal transitions.

source

DependencyTrees.ExplorationPolicy — Type.

ExplorationPolicy(k, p)

Simple exploration policy from Goldberg & Nivre, 2012. Returns true at rate p.

source