Trees

Dependency structure in natural language consists of directed relations between words in a sentence.

Simple API for building dependency trees:

DependencyTrees.DependencyTree — Type

DependencyTree

A rooted tree of dependency relations among the tokens of a sentence.

DependencyTree functions

DependencyTrees.arcs — Function

arcs(tree::DependencyTree)

Return a vector of (src, dst) tuples, representing all dependency arcs in tree.

Projectivity

DependencyTrees.isprojective — Function

isprojective(tree::DependencyTree)

True when a tree is projective, false when non-projective.

julia> tree = DependencyTree([("john", 2), ("saw", 0), ("a", 4), ("dog", 2), ("yesterday", 2), ("which", 7), ("was", 4), ("a", 9), ("yorkshire", 10), ("terrier", 7)])
   ┌─────────── ROOT
   │        ┌─► john
   └─►┌──┌──└── saw
      │  │  ┌─► a
   ┌──│  └─►└── dog
   │  └───────► yesterday
   │        ┌─► which
┌──└───────►└── was
│           ┌─► a
│        ┌─►└── yorkshire
└───────►└───── terrier

julia> isprojective(tree)
false

Tokens

DependencyTrees.Token — Type

Token(form, head=-1, label=nothing)

A token in a dependency tree.

Visualizing Trees

DependencyTrees.jl by default displays DependencyTrees by drawing arrows representing the dependency arcs. For example,