Graph Parsing

The GraphParsing submodule provides implementations of graph-based parsing algorighms.

julia> using DependencyTrees.GraphParsing

In graph-based dependency parsing, trees are built all at once, rather than greedily (one arc at a time). A set of possible parses for a sentence is represented by a square matrix of arc weights. The DependencyGraph type is used for this.

The DependencyTrees.GraphParsing module implements the following algorithms:

Chu-Liu/Edmonds

Dependency Graphs

DependencyTrees.GraphParsing.DependencyGraph — Type

DependencyGraph

todo

Graph Decoding Algorithms

Chu-Liu/Edmonds

DependencyTrees.GraphParsing.chu_liu_edmonds — Function

chu_liu_edmonds(G)

Decode a dependency tree using the Chu-Liu/Edmonds algorighm.