Transition Systems

Something here about transition parsing, I suppose.

Transition Systems

Arc-Standard

DependencyTrees.ArcStandard — Type.

ArcStandard()

Transition system for for Arc-Standard dependency parsing.

Transitions

Transition	Definition

LeftArc(l)	(σ|s1|s0, β, A) → (σ|s0, β, A ∪ (s0, l, s1))

RightArc(l)	(σ|s, b|β, A) → (σ, b|β, A ∪ (b, l, s))

Shift	(σ, b|β, A) → (σ|b, β, A)

Preconditions

Transition	Condition

LeftArc(l)	¬[s1 = 0], ¬∃k∃l'[(k, l', s1) ϵ A]

RightArc(l)	¬∃k∃l'[(k, l', s0) ϵ A]

See Nivre 2004.

source

Static oracle function:

DependencyTrees.static_oracle — Method.

static_oracle(cfg, gold_tree)

Static oracle for arc-standard dependency parsing. Closes over gold trees, mapping parser configurations to optimal transitions.

source

Arc-Eager

DependencyTrees.ArcEager — Type.

ArcEager()

Arc-Eager transition system for dependency parsing.

Transitions

Transition	Definition

LeftArc(l)	(σ|s, b|β, A) → (σ, b|β, A ∪ (b, l, s))

RightArc(l)	(σ|s, b|β, A) → (σ, b|β, A ∪ (b, l, s))

Reduce	(σ|s, β, A) → (σ, β, A)

Shift	(σ, b|β, A) → (σ|b, β, A)

Preconditions

Transition	Condition

LeftArc(l)	¬[s = 0], ¬∃k∃l'[(k, l', i) ϵ A]

RightArc(l)	¬∃k∃l'[(k, l', j) ϵ A]

Reduce	∃k∃l[(k, l, i) ϵ A]

References

Nivre 2003, Nivre 2008.

source

Static oracle functions

DependencyTrees.static_oracle — Method.

static_oracle(cfg, tree, transition)

Default static oracle function for arc-eager dependency parsing.

See Goldberg & Nivre 2012. (Also called Arc-Eager-Reduce in Qi & Manning 2017).

source

DependencyTrees.static_oracle_prefer_shift — Method.

static_oracle_prefer_shift(cfg, gold_tree, arc=untyped)

Static oracle for arc-eager dependency parsing. Similar to the "regular" static oracle, but always Shift when ambiguity is present.

See Qi & Manning 2017.

source

Dynamic oracle functions

DependencyTrees.dynamic_oracle — Method.

dynamic_oracle(t, cfg::ArgEagerConfig, tree)

Dynamic oracle function for arc-eager parsing.

For details, see Goldberg & Nivre 2012.

source

Arc-Hybrid

DependencyTrees.ArcHybrid — Type.

ArcHybrid()

Arc-Hybrid system for transition dependency parsing.

Described in Kuhlmann et al, 2011, Goldberg & Nivre, 2013.

source

Static oracle function:

DependencyTrees.static_oracle — Method.

static_oracle(::ArcHybrid, tree)

Static oracle for arc-hybrid dependency parsing. Closes over gold trees, mapping parser configurations to optimal transitions.

source

Dynamic oracle functions

DependencyTrees.dynamic_oracle — Method.

dynamic_oracle(t, cfg::ArgHybridConfig, tree)

Dynamic oracle function for arc-hybrid parsing.

For details, see Goldberg & Nivre, 2013.

source

Arc-Swift

DependencyTrees.ArcSwift — Type.

ArcSwift()

Parser configuration for arc-swift dependency parsing.

Described in Qi & Manning 2017.

source

Static oracle function:

DependencyTrees.static_oracle — Method.

static_oracle(::ArcSwiftConfig, graph)

Return a training oracle function which returns gold transition operations from a parser configuration with reference to graph.

Described in Qi & Manning 2017.

source

List-Based Non-Projective

DependencyTrees.ListBasedNonProjective — Type.

ListBasedNonProjective()

Transition system for list-based non-projective dependency parsing.

Described in Nivre 2008, "Algorithms for Deterministic Incremental Dependency Parsing."

source

Static oracle function:

DependencyTrees.static_oracle — Method.

static_oracle(::ListBasedNonProjectiveConfig, tree)

Return a training oracle function which returns gold transition operations from a parser configuration with reference to graph.

source

Misc.

For stack-and-buffer transition systems (Arc-Eager, Arc-Standard, Arc-Hybrid, and Arc-Swift), DependencyTrees.jl implements functions for getting the tokens from the stack and buffer in a safe way:

DependencyTrees.stacktoken — Function.

stacktoken(cfg, i)

Return the token at stack index i (starting at 1).

source

DependencyTrees.buffertoken — Function.

buffertoken(cfg, i)

Return the token at buffer index i (starting at 1).

source