Treebanks

A Treebank is a corpus of dependency-annotated sentences in one or more files.

DependencyTrees.Treebank — Type

Treebank

A reader for a file with annotated parse trees.

Iterating over a treebank reads sentences one at a time:

treebank = Treebank("data/news.conll", conllu)

for tree in treebank
    # ...
end

tree = first(treebank)

# output
┌────────────── ROOT
│           ┌─► Economic
│        ┌─►└── news
└─►┌──┌──└───── had
   │  │     ┌─► little
   │  └─►┌──└── effect
   │  ┌──└────► on
   │  │     ┌─► financial
   │  └────►└── markets
   └──────────► .

Supported Formats

CoNLL-U

DependencyTrees.conllu — Function

conllu(text)

Read a token (in a dependency tree) from CoNLL-U format.

CoNLL-X

conllx ```