Trees

Dependency structure in natural language consists of directed relations between words in a sentence.

Simple API for building dependency trees:

DependencyTrees.deptree — Function

deptree(read_token, xs)

Create a DependencyTree by calling read_token on each of xs.

DependencyTrees.deptoken — Function

deptoken(form, head=-1, label=nothing)

Create a token in a dependency tree.

DependencyTrees.DependencyTree — Type

DependencyTree

A rooted tree of dependency relations among the tokens of a sentence.