Trees

Trees

This is where I write about trees.

Dependency structure in natural language consists of directed relations between words in a sentence.

DependencyTrees.DependencyTree — Type.

DependencyTree

A natural language sentence annotated with dependency structure.

source

DependencyTrees.check_tree — Function.

check_tree(g, check_single_head=true, check_has_root=true, check_projective=false)

Ensure the well-formedness of the dependency graph g, throwing an error if g is not well-formed.

source

Types of tokens

DependencyTrees.UntypedDependency — Type.

UntypedDependency(id, form, head)

Represents a simple untyped (head –> dependent) relation between two words in a sentence.

source

DependencyTrees.TypedDependency — Type.

TypedDependency(id, form, deprel, head)

Represents a typed (head -[deprel]-> dependent) relation between two words in a sentence. id is the index of the word form in the sentence (starting at 1, with 0 meaning the root node). deprel is the dependency relation between this word and its head).

source

DependencyTrees.CoNLLU — Type.

CoNLLU

CoNLLU token in a dependency tree.

See https://universaldependencies.org/format.html.

source